requirements.txt
# Core dependencies
firebase-admin==6.6.0
langchain==0.1.20
langchain-core==0.1.52
langchain-openai==0.0.8
langchain-community==0.0.38
openai==1.66.2
python-dotenv==1.0.1
uvicorn==0.27.1
fastapi==0.110.0
aiohttp==3.9.3
pydantic-settings==2.8.1
